revert to the old way of updating the matrix
the old way is better because if there is more than one level in critical state, as soon as the level displayed in the matrix gets normalized, its going to update the matrix to display the other level still critical. This way feels better even if its constantly updating the error matrix. I might make a better logic later, this will do for now.